backend/feature1/requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
requests==2.28.1
beautifulsoup4==4.11.1
selenium==4.12.0
webdriver-manager==4.0.1
functions-framework==3.4.0
google-cloud-firestore==2.11.1
dnspython==2.2.1
python-whois==0.8.0